refactor lookup tables into a package (#358)
