--- docs/source/api_reference/framework/lookup.rst
@@ -1,1 +1,0 @@
-.. automodule:: vivarium.framework.lookup
