+++ docs/source/api_reference/framework/lookup/index.rst
@@ -1,0 +1,7 @@
+.. automodule:: vivarium.framework.lookup
+
+.. toctree::
+   :maxdepth: 2
+   :glob:
+
+   *
