+++ docs/source/api_reference/framework/lookup/interpolation.rst
@@ -1,0 +1,1 @@
+.. automodule:: vivarium.framework.lookup.interpolation
