+++ docs/source/api_reference/framework/lookup/manager.rst
@@ -1,0 +1,1 @@
+.. automodule:: vivarium.framework.lookup.manager
