--- docs/source/api_reference/interpolation.rst
@@ -1,1 +1,0 @@
-.. automodule:: vivarium.interpolation
