+++ src/vivarium/framework/lookup/__init__.py
@@ -1,0 +1,6 @@
+from vivarium.framework.lookup.manager import LookupTableInterface, LookupTableManager
+from vivarium.framework.lookup.table import (
+    LookupTable,
+    LookupTableData,
+    validate_parameters,
+)
